Sign up page redesign + functionality tested
