+++ client/src/App.js
@@ -9,14 +9,14 @@ import AOS from 'aos';
 import Home from './pages/Home';
 import NoMatch from './pages/NoMatch';
 import Dashboard from './pages/Dashboard';
-import SignUp from './pages/SignUp';
-import Login from './pages/Login';
+import SignUp from './pages/SignUpOld';
+import Login from './pages/LoginOld';
 import AddHappyHour from './pages/AddHappyHour';
 import Feed from './pages/Feed';
 import HHPost from './pages/HHPost'
 import Home2 from './pages/Home2'
-import SignIn2 from './pages/SignIn2';
-import SignUp2 from './pages/SignUp2'
+import SignIn from './pages/SignIn';
+import SignUp2 from './pages/SignUp'
 
 
 
@@ -55,7 +55,7 @@ const App = () => {
 			<Route
 				path='/signup'
 				element={
-					<SignUp2 />
+					<SignUp />
 				}
 			/>
 			<Route
@@ -67,7 +67,7 @@ const App = () => {
 			<Route
 				path='/signin'
 				element={
-					<SignIn2 />
+					<SignIn />
 				}
 			/>
 			<Route
